Mock storage during development.
