+++ lib/packages.js
@@ -1,5 +1,3 @@
-"use strict";
-
 var PackageModel = require('./models').PackageModel
 	, error = require('./error')
 	, users = require('./users')
@@ -7,19 +5,11 @@ var PackageModel = require('./models').PackageModel
 	, _ = require('underscore')
 	//, search = require('./search')
 	, mongoose = require('mongoose')
-	, amazonS3 = require('awssum-amazon-s3')
 	, UserModel = require('../lib/models').UserModel
-  , fs = require('fs')
-  , fmt = require('fmt')
-	, crypto = require('crypto');
+    , fs = require('fs')
+	, crypto = require('crypto')
+    , storage = require('./storage');
 
-var s3 = new amazonS3.S3({
-    'accessKeyId'     : process.env.AWSAccessKeyId,
-    'secretAccessKey' : process.env.AWSSecretKey,
-    'region'          : amazonS3.US_EAST_1,
-});
-
-var S3_BUCKET_NAME = process.env.NODE_ENV != "production" ? "greg-pkgs-testing" : "greg-pkgs-prod";
 
 // Array Remove - By John Resig (MIT Licensed)
 Array.prototype.remove = function(from, to) {
@@ -675,40 +665,16 @@ function _validate_new_pkg_version(req, pkg_data, callback) {
 					return;
 			  }
 
-			  var bodyStream = fs.createReadStream( req.files.pkg.path );
-
-			  var objectName = exports.guid() + req.files.pkg.name;
-
-	      var options = {
-	        BucketName    : S3_BUCKET_NAME,
-	        ObjectName    : objectName,
-	        ContentLength : req.files.pkg.size,
-	        Body          : bodyStream
-	      };
-
-			  try {
-			    s3.PutObject(options, function(err, data) {
-			      fmt.dump(err, 'err');
-			      fmt.dump(data, 'data');
-						
-						if (err){
-							console.error( err )
-							return outer_callback( error.fail('There was an error saving the package contents.') );
-			      }
-
-			      // get the url
-			      pkg_data.url = "https://s3.amazonaws.com/" + S3_BUCKET_NAME + "/" + objectName;
-			      return outer_callback(null, users);
-
-			    });
-			  } catch (e) {
-			    return outer_callback( error.fail('There was an exception trying to save the package contents.') );
-			  }
+			  try{
+                  storage.upload(req, pkg_data, exports.guid());
+              }catch(err){
+                  outer_callback(err);
+              }
+              
+              outer_callback(null, users);
 
 			});
-			
 		}
-
 	], function(err) { // closing waterfall 
 
 		if (err) {
@@ -874,41 +840,16 @@ function _validate_new_pkg(req, pkg_data, callback) {
 					return;
 			  }
 
-			  var bodyStream = fs.createReadStream( req.files.pkg.path );
-
-			  var objectName = exports.guid() + req.files.pkg.name;
-
-		      var options = {
-		        BucketName    : S3_BUCKET_NAME,
-		        ObjectName    : objectName,
-		        ContentLength : req.files.pkg.size,
-		        Body          : bodyStream
-		      };
-
-			  try {
-			    s3.PutObject(options, function(err, data) {
-			      fmt.dump(err, 'err');
-			      fmt.dump(data, 'data');
-
-
-			      if (err){
-			      	console.error( err )
-							return outer_callback( error.fail('There was an error saving the package contents.') );
-			      }
-
-			      // get the url
-			      pkg_data.url = "https://s3.amazonaws.com/" + S3_BUCKET_NAME + "/" + objectName;
-			      outer_callback(null, users);
-
-			    });
-			  } catch (e) {
-			    return outer_callback( error.fail('There was an exception trying to save the package contents.') );
-			  }
+			  try{
+                  storage.upload(req, pkg_data, exports.guid());
+              }catch(err){
+                  outer_callback(err);
+              }
+              
+              outer_callback(null, users);
 
 			});
-
 		}
-
 	// closing waterfall
 	], function(err) {
 		if (err) {
