+++ test/mock_bucket/.gitignore
@@ -1,0 +1,1 @@
+*.zip
