+++ test/test_rest.js
@@ -40,6 +40,7 @@ var groups = ['global', 'the cool group', 'cool kidz', 'adsk'];
 var pkg_datas = [];
 var ds_pkg_datas = [];
 
+
 // This code attempts to insert a large number of elements into the package manager database.
 // It has the side effect of actually populating the database with these elements, so be wary of
 // when and how you use it.
